Role-based auth: migrations, models, middleware and login redirect
- Add roles, permissions, role_permissions, contact_emergency migrations
- Fix role_id as unsignedBigInteger with FK to roles table
- Add Role and Permission Eloquent models with relationships
- Add role() BelongsTo relationship to User model
- Create RoleMiddleware checking role via roles.name_role
- Register 'role' alias in Kernel
- Redirect after login based on user role (admin/employe/default)
- Create EmployeController and employe dashboard view/route

+++ database/migrations/2026_04_06_022504_update_users_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->string('social_segurity_number', 20)->after('curp')->nullable()->unique();
             $table->date('birthdate')->after('social_segurity_number')->nullable();
             $table->integer('id_contact_emergency')->after('birthdate')->nullable();
-            $table->integer('role_id')->after('id_contact_emergency');
+            $table->unsignedBigInteger('role_id')->after('id_contact_emergency');
         });
     }
     /**
